@@ -6,36 +6,60 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build",
+    "build": "webpack --mode production",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
     "@angular/animations": "^16.1.0-rc.0",
     "@angular/common": "^16.1.0-rc.0",
     "@angular/compiler": "^16.1.0-rc.0",
-    "@angular/core": "^16.1.0-rc.0",
+    "@angular/core": "^16.2.9",
     "@angular/forms": "^16.1.0-rc.0",
     "@angular/platform-browser": "^16.1.0-rc.0",
     "@angular/platform-browser-dynamic": "^16.1.0-rc.0",
     "@angular/router": "^16.1.0-rc.0",
+    "@types/prismjs": "^1.26.3",
+    "angular-highcharts": "^16.0.0",
     "angular-in-memory-web-api": "~0.16.0",
     "bootstrap": "^5.3.2",
     "echarts": "^5.4.3",
+    "highcharts": "^10.3.3",
+    "highcharts-angular": "^3.1.2",
+    "jquery": "^3.7.1",
+    "markdown-it": "^13.0.2",
+    "marked": "^11.0.1",
     "ngx-echarts": "^16.1.1",
+    "prismjs": "^1.29.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "^0.13.0"
+    "zone.js": "^0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.0-rc.0",
+    "@angular-devkit/build-angular": "^16.2.7",
+    "@angular-eslint/builder": "16.3.1",
+    "@angular-eslint/eslint-plugin": "16.3.1",
+    "@angular-eslint/eslint-plugin-template": "16.3.1",
+    "@angular-eslint/schematics": "16.3.1",
+    "@angular-eslint/template-parser": "16.3.1",
     "@angular/cli": "^16.1.0-rc.0",
     "@angular/compiler-cli": "^16.1.0-rc.0",
+    "@eslint/js": "^9.3.0",
     "@types/jasmine": "~4.3.0",
+    "@types/jquery": "^3.5.25",
+    "@types/markdown-it": "^13.0.7",
+    "@types/marked": "^6.0.0",
     "@types/node": "^16.11.35",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
     "copyfiles": "^2.4.1",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.51.0",
+    "eslint-config-prettier": "^9.1.0",
+    "globals": "^15.3.0",
     "jasmine-core": "~4.6.0",
     "jasmine-marbles": "~0.9.2",
     "jasmine-spec-reporter": "~7.0.0",
@@ -44,7 +68,9 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "protractor": "~7.0.0",
+    "prettier": "^3.2.5",
+    "protractor": "^7.0.0",
+    "style-loader": "^3.3.3",
     "ts-node": "~10.9.0",
     "typescript": "~4.9.3"
   }
