@@ -1,19 +1,20 @@
 /*
-*  Protractor support is deprecated in Angular.
-*  Protractor is used in this example for compatibility with Angular documentation tools.
-*/
-import { bootstrapApplication,provideProtractorTestingSupport } from '@angular/platform-browser';
+ *  Protractor support is deprecated in Angular.
+ *  Protractor is used in this example for compatibility with Angular documentation tools.
+ */
+import {
+  bootstrapApplication,
+  provideProtractorTestingSupport,
+} from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
 import { provideRouter } from '@angular/router';
 import routeConfig from './app/routes';
-import {HttpClientModule, provideHttpClient} from "@angular/common/http";
+import { HttpClientModule, provideHttpClient } from '@angular/common/http';
 
-bootstrapApplication(AppComponent,
-  {
-    providers: [
-      provideProtractorTestingSupport(),
-      provideRouter(routeConfig),
-      provideHttpClient(),
-    ]
-  }
-).catch(err => console.error(err));
+bootstrapApplication(AppComponent, {
+  providers: [
+    provideProtractorTestingSupport(),
+    provideRouter(routeConfig),
+    provideHttpClient(),
+  ],
+}).catch(err => console.error(err));
