@@ -2,5 +2,6 @@ export class Site {
   constructor(
     public job_name: string,
     public city_name: string,
-  ) {  }
+    public wor_exp: string
+  ) {}
 }
